fix 4.1

+++ tasks/stl/Konstantinov_I_Sort_Batcher/src/ops_stl.cpp
@@ -115,7 +115,7 @@ void RadixSorted(std::vector<double>& arr) {
   }
   size_t n = arr.size();
   std::vector<uint64_t> keys(n);
-  const int thread_count = static_cast<int>(ppc::util::GetPPCNumThreads());
+  const int thread_count = ppc::util::GetPPCNumThreads();
 
   ParallelConvertToKeys(arr, keys, thread_count);
